images: Rework the part regarding the rauc bundle
Before we used to look for the rauc bundle and casync store in a
directory named "rauc". This was changed in the image recipe, now the
bundle and the store live at the same level as the other artifacts, and
also bear the image names, with the `.raucb` extension for the bundle,
and `.castr` extension for the store.

This is more consistent with the naming convention already in use for
other build artifacts.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>