server: Measure the time needed to walk the image directory
An image is made of several build artifacts. Among them, the casync
store, which contains thousands of files and directories. Around 100k
entries, to give you an idea.

I'm starting to suspect that this will be a problem. This will
tremendously slow down the server startup, as the server walks the
image hierarchy, hence will have to walk the casync store. Once we start
to have many images, and many casync stores, we will have to find
another solution, cause I'm afraid it will just be too slow.

But let's do things in order: measure first, optimize later.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>