client: Add a ImagesUrl setting
I'm not sure what's the plan about the server where images actually
live. It does not need to be the same url as the QueryUrl. And at the
moment, due to the setup I use for my tests, it can't be the same url.

So for now, let's just have a mandatory setting in the client
configuration, it solves my current issue, and it has the merit to be
explicit, rather than some magic hidden somewhere in the code.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>