---- Complete rewrite ----
This is almost a complete rewrite of the code.

Mostly, this second iteration improves the overall design, with better
abstractions, cleaner code, and removes any weird code here and there.
The code is now quite clean, albeit far from perfect.

And now, a few words on dataclasses...

One of the main changes in this rewrite is to use python dataclasses, a
feature from python 3.7. The idea was to reduce the boilerplate code
needed to go from JSON to dictionary to object, and the other way
around. The result is quite disappointing.

First, dataclasses provide `asdict()` to serialize an object to a
dictionary, but do not provide `fromdict()` to go from a dictionary to
an object. So it solves only half of the problem.

Additionally, `asdict()` is not super flexible, because it recurses on
the object it's called on, and there's no way to hook yourself anywhere.
So for any objects that is not suitable to be serialized 'as is', you're
stuck, all you can do is to make it a normal class, not a dataclass, so
that the asdict() recursion stops there.

  (the concrete example at this point of the code is the BuildID class,
  that I want to serialize as a string such as `20190214.1`, not as a
  dict such as `{ 'date': datetime.date(2019, 02, 14), 'incr': 1 }`)

Past this point, asdict() gives you a dictionary, which contains a mix
of builtin types, and objects.  These objects are not suitable to be
serialized to JSON, so we need an additional hook, by subclassing
JSONEncoder and handling our objects explicitly there.

Past this point, we have the serialization worked out, and handled
automatically.

Now comes deserialization: once again, we need to hook ourselves
somewhere to handle our custom objects, which can be done by subclassing
JSONDecoder to start with. I didn't try it, so I'm not sure if it's
enough, but I suspect there would be more issues here and there.

Anyway.

So the bottom-line is: automatic serialization/deserialization that does
"just the right thing" just doesn't exist, and dataclasses don't help
much, unless you REALLY have absolutely NO object that requires a custom
(de)serialization. That's doesn't go far.

I could have tried harder to get it all automated, but the risk is to
end up with a solution that works today, and doesn't work tomorrow
because of a new type/object that require new custom handling, and where
the automated solution just can't work.

For this reason, in this new version of the code, there's still a lot of
boilerplate methods named `from_dict()` and `to_dict()`. It's not that
bad, because this code is dumb, hence cheap to maintain, and it's
flexible as we can handle pretty much everything explicitely. Good
enough for now, let's live with that.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>