server: Do not provide a default config file anymore
A default config file is good when there's only one instance of a daemon
that should be running. But the update server is more like an http
server, where you have a list of sites you want to serve. In this
context, a default config file doesn't really make sense.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>