Renaming: steamosupdate > steamosatomupd (python modules)
'steamosupdate' is too generic a name, and it doesn't say enough about
what we're talking about. In the long run it might just be confusing for
everyone.

'steamosatomicupdate' is a much better name, albeit a bit long. In the
context of python module names, and according to PEP 0008 [1], let's
shorten it a bit to 'steamosatomupd'

[1]: https://www.python.org/dev/peps/pep-0008/#package-and-module-names

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>