client: Add a hack regarding unfreezing rootfs
This line doesn't belong here. We might have frozen the rootfs at some
point, but this is done in the RAUC pre-install script, hence unfreezing
shoud be done in the post-install script. Except that for now, if the
update fails, RAUC doesn't run any post-install script, and there's no
way to run a script on failure.

So let's put this here for now, until I come up with a better solution.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>