Move want-unstable from client to server side
This brings us closer to our ideal: having a client as thin and dumb as
possible, and having most of the code and logic server-side instead.

Additionally, the code is actually simpler server-side, as we don't
have to carry a want_unstable parameter around.

And the last bonus is that it also simplifies writing unit tests. Before
it was a bit awkward to test the want-unstable thing, and that's exactly
why it was not tested, and left as a TODO. Now that we moved it
server-side, adding a test in the testuite is straightforward.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>