client: print the tips to see real progress
RAUC is a client/daemon update service.

The script steamos-atomupd-client runs the client part of RAUC that
tells to the daemon part of RAUC to start an update.

RAUC has its own progress status; but it is not appropriate with the use
of casync. RAUC does not forward logs from casync and the update takes
several minutes without any progression.

The commit prints the tips to see the RAUC progression.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>