client: Move default manifest file to /etc/steamos-atomupd/manifest.json
For testing it can be convenient to have that file in a non-read-only
area; /etc is read-write.

Furthermore, os-release has two locations for the file that contain its
identification data: /etc/os-release and /usr/lib/os-release.

Note: /etc/os-release is be a relative symlink to /usr/lib/os-release;
lets imitate the behavior at the distribution level.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>