client: Replace the option --progress by --quiet
The process takes several dozen of minutes to completed and does not
output anything to stdout.

The progression is the best feedback to output to the user since the
thread is expected to exit properly since the commits 64954e1 (client:
Use multiprocessing for progress) and d936465 (client: Exit the thread
loop if RAUC fails).

This replaces the option --progress by --quiet that defaults to output
the progression to stdout unless the option --quiet is set.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>