server: Return the HTTP 400 status code if the request is malformed
From the datastructures.py documentation about the MultiDict class
(`request.args`):

"From Werkzeug 0.3 onwards, the `KeyError` raised by this class is also
a subclass of the :exc:`~exceptions.BadRequest` HTTP exception and will
render a page for a ``400 BAD REQUEST`` if caught in a catch-all for
HTTP exceptions."

If we do not catch this `KeyError`, when we try to access a member that
is not available, Flask will automatically return the HTTP status code
400.
However if an expected arguments have unexpected values we raise a
`ValueError` and Flask will return the HTTP status code 500, that is not
what this server was supposed to do.

Instead of relying on this implicit conversion it's better to be
explicit about what we want to return when the arguments are malformed
(missing parameters or with unexpected values).

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>