client: Do not remove the update info in --query-only mode
The `--query-only` mode is supposed to bail out before applying an
update.
As explained in the documentation `client.md`, it could be used for
example to fetch information about the available updates, without
downloading them immediately, and the resulting JSON file will be stored
in a well-known location.

For this reason we shouldn't delete the downloaded JSON file when using
the `--query-only` argument.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>