CI: Build a steamos-atomupd Docker image
We build on the following events:

* push a tag to the repository
* push a branch to the repository
* timed, triggered or manually-requested build from any branch

We do not redundantly build for merge requests.

For tags, the image is tagged as something like
registry.gitlab.steamos.cloud/steam/steamos-atomupd:0.20211124.0 and
also as ...:latest, and is pushed to the registry.

For unreleased commits on the default branch master (or main), the
image is tagged as something like ...:0.20190724.0-48-g2aad2e3
and also as ...:master, and is pushed to the registry.

For unmerged commits on other branches, for example wip/smcv/ci,
the image is built but not pushed.

Signed-off-by: Simon McVittie <smcv@collabora.com>