Also write json files with less information known in staticServer
Because candidates are mapped in ImagePool as:
Product
  Release
    Variant
      Arch: [ candidate1, candidate2 ]

As soon as we move up the json tree to where variant is not known
we get no data at all.

I think to fix this we need to change to use the same mapping as
we are writing the parameter paths. i.e.

Product
  Arch
    Release
      Variant: [ candidate1, candidate2 ]

Then when getting possible updates for unknown variant return
all candidates for all variants with a given product, arch, release