client: Add option to choose a specific update version
Currently the major updates have an higher priority compared to the
minor updates, without a way to set your preference.

I decided to add an option that allows you to give in input a
specific version ID, instead of just a flag to force only major/minor
updates, because this will avoid unexpected surprises and will align it
with the API that is currently provided in `atomupd-daemon`
(https://gitlab.steamos.cloud/steam/atomupd-daemon/-/merge_requests/2).
For example if you first run with `--query-only`, the second time you
try to apply an update, unless you stored the update json file
somewhere, you might end up installing an unexpected update because the
server may return a different list of available updates.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>