client: Add ability to set a custom variant as a launch argument
Currently the "variant" value is used to switch between different update
channels, e.g. from "steamdeck" to "steamdeck-beta".

Because this value is taken from the manifest file, changing the variant
is not a straight forward process.

By introducing the `--variant` argument we are able to make the whole
process easier.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>