Add estimated_size as an update candidate optional key
With Desync we now have the ability to estimate a download size for the
updates.

With this commit we add the key "estimated_size" to hold this
information.
At the moment the server is not expected to calculate this value, so
it's up to the client to locally estimate the download size for a
particular update while using its active partition as a seed.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>