client: Refactor do_update()
We already ensure that `images_url` ends with a forward slash in run().
For this reason there is no need to duplicate this check in do_update().

Also just passing an update URL makes it easier to launch updates when
we already have a complete address instead of the separated images and
update path.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>