client: Add support for QueryUrl as a fallback
Recently this client helper has been updated to support the new static
server configuration, using the MetaUrl option.

However the static server is not yet used in production and this switch
is preventing to deploy further updates in "steamos-atomupd" until the
switch to the static server is actually completed.

With this commit we add back the support for the older `QueryUrl` as a
fallback, to also handle the previous smart servers requests and ease
this transaction period.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>