client: Get rootfs path from the "booted" slot
The rootfs device path was taken from the "boot_primary" entry in RAUC
status output.

While this was correct most of the time, if we already applied an update
and had a pending reboot, the path would point to the wrong rootfs path.
Because when we apply an update we change the bootloader selected image,
and that is also reflected in RAUC's "boot_primary".

Instead, with this commit, we take the rootfs path by looking at the
slot that has the "booted" state, i.e. the slot that is currently
active.

Partially addresses:
https://gitlab.steamos.cloud/jupiter/tasks/-/issues/473

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>