client: Try to prevent update loops
If the server proposes us to "update" to the same current version, we
will end up in an infinite update loop.

We should instead gracefully handle that situation.

This will ease the transition from the old smart server to the new
static server, because we will be able to point the old query URL to a
static file, without having to worry about the version that is in use by
those devices in the transition phase.

Fixes: https://gitlab.steamos.cloud/jupiter/tasks/-/issues/472

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>