staticserver: Return an empty JSON object {} when there are no updates
The expected reply, when no updates are available, is the empty JSON
object (i.e. {}) and not the string "{}".
https://gitlab.steamos.cloud/steam/steamos-atomupd/-/blob/2aad2e3e/doc/client.md

The JSON is parsed in
https://gitlab.steamos.cloud/steam/steamos-atomupd/-/blob/22847a03/steamosatomupd/client.py#L653
and, parsing a file that contains `{}` will result in `update_data`
being a dict with zero elements. Instead parsing a file that has "{}"
will assign `update_data` a str "{}".
That will break the assumption of using a dict in `from_dict()`.

We are currently not using the static server in production, so there is
no risk of regressions.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>