client: Support parsing the new detailed Desync progress
Initially Desync was just printing in output the progress percentage of
the assembling phase, leaving behind the preparation steps.

With the PR #217, now Desync prefixes the progress with the name of the
step it is currently executing, and also prints a separate progress
percentage for the validation and chunking phases.

With this commit we add support to parsing this new syntax, while
retaining the older one too, to ease the transaction period.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>