client: Add support for letting Desync regenerate the index itself
With the new argument `--regenerate-invalid-seeds`, we are able to use a
seed index that might even be invalid. Desync will validate the index
and regenerate the seed itself, if needed.

This allows us to speed up the preparation phase.

The index is store since the MR
https://gitlab.steamos.cloud/steam/steamos-customizations/-/merge_requests/142
However the symlink to the active rootfs will still be handled by
client.py, to ensure that we are pointing to the correct partition.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>