client: Ensure the rootfs is in RO mode while applying an update
A user could still re-enable the RW mode while the update is in
progress. However this is not a severe problem.
Actively writing to the rootfs while applying an update could result in
an aborted update (Desync) or an update that requires to download more
chunks than what was initially expected (Casync).

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>