client: Support the RAUC fallback configuration
When a system update fails multiple times we change RAUC service
configuration to the fallback one, if available.

That will allows us to have a secondary fallback configuration that has
different/more conservative options.

The failed attempts are store in the runtime directory
"/run/steamos-atomupd".
Because this directory is expected to be mounted in tmpfs, after a
reboot, the counter of failed updates will be automatically erased.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>