Add ability to consider more stable variants when checking for updates
With the new config option 'VariantsGroupOrder' we can now instruct the
server to take into consideration the more stable variants too, when it
needs to decide if an update is available or not.

The elements in 'VariantsGroupOrder' are ordered from the most stable to
the least stable variant. Additionally they must be all related to the
same "class"/"group". Because for example it wouldn't make sense to
offer an holo update if the device is currently running steamdeck-beta.

E.g. with this new option if we have `steamdeck` and `steamdeck-beta`
variants, and the `steamdeck` has a build that is newer than the
`-beta`, the server will propose an update to the latest `steamdeck`
even if you asked for updates in `steamdeck-beta`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>