Allow branch switching while searching for updates
Currently Jupiter has multiple variants (aka branches). E.g. steamdeck,
steamdeck-beta and steamdeck-main.

The users are expected to be able to freely switch between them.

With this commit we add support for images to ask for updates in a
variant that is different from the one in use.

The current implementation assumes that a buildid greater than a
checkpoint is expected to be compatible with it, even for checkpoints of
different branches.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>