staticserver: Replace prints with logging
The messages that we were printing were just report of the events that
were occurring during the execution.
For this type type of messages the recommended tool is `logging`.

The `Jsonresult` prints have been replaced with `log.debug` because they
are incredibly verbose and with a non trivial amount of update images
they'll just flood the output, potentially covering other more useful
messages.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>