staticserver: Check the eventual older JSON before replacing it
If the meta update JSON file that we are going to write is already
present and equal, we log that nothing changed and we move on to the
next file.
Instead, if the new JSON file is different, before replacing it, we log
the differences between the two files.

This will allow a quick manual check to verify that the static server
worked as expected.

Example output:
INFO:staticserver.py:147: Replacing "s/a/3.1.0/sd/20220301.100.json":
- "buildid": "20220301.1",
?                    ^ ^
+ "buildid": "20220303.2",
?                    ^ ^
- "estimated_size": 468332221,
?                   ^^   ---
+ "estimated_size": 508732331,
?                   ^^ +++
- "update_path": "s/h/3.0/a/steamos-holo-20220301.1-3.0-amd64-sd.raucb"
?                                               ^ ^
+ "update_path": "s/h/3.0/a/steamos-holo-20220303.2-3.0-amd64-sd.raucb"
?                                               ^ ^

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>