client: remove parsing deprecated "Slot" entry in RAUC output
Since RAUC v1.6, the message "Slot needs to be updated with [...]" has
been removed.
https://github.com/rauc/rauc/commit/40cd43bf0d26a6e26667ae5c93c3a100ac1ee07b

We were using this line as an entry point to ensure that the seeding and
downloading were related to the rootfs image.

We could have replaced it by parsing the "Updating [...]" entry instead,
but because they are just log messages and not a stable API, we should
just parse the messages that are strictly necessary.
The rootfs is the only slot we currently update and that's not likely to
change without a complete redesign of our atomic updates. So we can
simply skip the "slot" check and go straight to the update progress.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>