build-deps: Use sha1 instead of tag names
We are cloning RAUC from its official Github repository.

Given that this is a public repository that we don't control, it's
generally safer to use the hash instead of a tag name. This guards us
from installing an unexpected version in case the tag got maliciously
replaced by an attacker.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>