utils: Take into account the chunks compression rate
The chunks are stored as compressed archives. This means that the "size
not in seed" provided by Desync, is actually bigger than the actual
download size required for a particular update.

By manually downloading several updates and comparing the estimated size
with the actual used bandwidth, we calculated that the average
compression rate of the chunks is about 1.33-1.50.

With this knowledge, we can adjust the estimation to take into account
the compression rate.
Of course this is not an exact figure, but it's a good enough
approximation to give users an estimated size that is in the ballpark of
the real value.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>