Add option to skip undesired images
When we want to skip an image from being considered in the update pool,
now we can use the "skip" option.
This new option gives us the ability to briefly explain why an image
should be skipped (e.g. because it has serious bugs) and it's also
understood by the server that can correctly handle it.

If a client is using an image that is marked to be skipped, the server
will attempt to propose either an upgrade, whether possible, or a
downgrade.

Alternative methods to remove an image from the candidates pool, like
literally deleting the image manifest or editing the manifest to
invalidate it, should be avoided at all cost, unless you really
understand the consequences.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>