Use pyinotify to watch for a trigger file to parse data.
Watch for a specific trigger file in each subfolder of PoolDir
to be created or modified.
When that file is created parse all data using the already
parsed config file.
Added a -s|--run-single mode to not daemonize and run one parse
to use in the static server test.
--run-single is the default.
Use -r or --run-daemon to run as a daemon. Otherwise
only do one parse round and exit.