imagepool: Fix fallback updates by adding proper support for them
In some edge cases we were proposing unexpected updates for the generic
variant fallback JSON, i.e. update files for images whose buildid is
unexpected/borked, as highlighted in the changes introduced by this
commit.

So far, for the fallbacks, we created a fake image with the buildid
19000101, to simulate a generic old image. While that was an effective
trick for unversioned images, it generated unexpected results when we
started to mix versioned and unversioned images.
This is due to the fact that we were comparing an image with a very old
buildid but with a recent version.
E.g. in the "releases-and-snaps" test we had an image (steamdeck-beta,
3.1, 20220301.100) and we wanted to generate a generic fallback that
switched to the `steamdeck` variant. To do that we edited that image
into (steamdeck-beta, 3.1, 19000101.1).
But then the server returned that our update was (steamdeck, snapshot,
20220225.1), instead of the expected (steamdeck, 3.0, 20220303.2).
This was caused by the fact that (steamdeck-beta, 3.1, 19000101.1) is in
fact *newer* than (steamdeck, 3.0, 20220225.1), because the first
element of comparison is the version. And the buildid is compared only
after that.

To properly fix that, we introduce a new function parameter called
`unexpected_buildid`. With this boolean we can signal the imagepool that
we are interested in generating a fallback JSON update, without having
to rely on manually altering the buildid of a genuine image.

Given that in a fallback we can't know for sure which buildid a
particular client is actually using, the logic here assumes that the
client already has all the checkpoints that were introduced in the
earlier versions.
For example, if in version 3.6 there was a checkpoint, and you are
asking for the generic fallback `3.7.0/steamdeck.json`, we assume that
the checkpoint in 3.6 doesn't need to be reinstalled again.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>