Do not consider more stable images if they are still snapshots
With snapshots we can't really order images from different variants,
because only their buildid can be used as a comparator.

This means that if we use the VariantsOrder with snapshots, we might end
up with an hotfix in the stable variant being preferred over the
development main images. Only because its buildid is newer.

With this commit we add a special case for snapshot images in order to
avoid considering them when evaluating the VariantsOrder.
In this way we can start to use the new VariantsOrder right away,
without having to wait for versioned images to propagate in every
channel.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>