Add UpdateType Enum to select the type of update
Create an Enum to group together the type of updates that we can search
for.

This helps to reduce the number of parameters that we need to pass and
makes it clearer that they are exclusive, e.g. it didn't make sense to
ask for an update that was both "forced" and with an "unexpected
buildid".

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>