imagepool: Add TODO regarding the checkpoints
The checkpoint logic needs to be revised because in our code we are
making assumptions that are not guaranteed, i.e. ordered candidates, and
also because we can't do the right thing for the fallback updates
without having additional image info from the client.

For example in `3.6/steamdeck.json` we only know that the client is
using the version 3.6, but we don't know if, and how many, checkpoints
it already installed.

The discussion about how this should be improved is in:
https://gitlab.steamos.cloud/jupiter/tasks/-/issues/912

Until this gets properly addressed, we include an inline TODO.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>