imagepool: Sort candidates before picking the correct updates
When we have multiple checkpoints, the implementation in
`_get_update_candidate()` implicitly required the candidates to be
sorted. Otherwise, we could have ended up proposing two checkpoints in
the wrong order, or picking just a newer checkpoint while not including
the previous ones.

With this commit we guarantee that the candidates are actually sorted in
ascending order and we revise `_get_update_candidate()` to explicitly
take that into account.

One assumption that we still make is that all variants are expected to
have equivalent checkpoints.
E.g. this means that if rel has two checkpoints, we expect that beta
also has two checkpoints that make the exact same breaking changes.
The discussion about how to handle and sort checkpoints can be found in
jupiter/tasks#912.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>