Add option to request the second last update
Add a generic fallback JSON that points to the penultimate image.

When a client is running the latest image available, this new second
last info can be used by `steamos-reset` to pick a different valid
image.
Ref: https://gitlab.steamos.cloud/jupiter/tasks/-/issues/701

If we have checkpoints, we don't propose second last updates because we
currently don't have the ability to verify if a client would be able to
apply said update or not.
Ref: https://gitlab.steamos.cloud/jupiter/tasks/-/issues/912

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>