tests: Programmatically create the server config files
With our increasing number of tests we are reaching the point where we
are polluting the "examples" directory with a lot of similar files.

Instead of shipping every single variation of those server config files,
we generate them directly in the Python tests.

The only config file that we kept is `server-releases.conf`,
which could be used as a reference file for the real server
configuration.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>