tests: Remove same image test
Having two images with the same version and buildid, while only one is
marked as a checkpoint is not a use case we support.

Remove this test because it doesn't represent something that should ever
happen in reality.

An automated test, to ensure this never happens in production, will be
added in a future commit.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>