Skip the "checkpoint" property when its value is not true
Checkpoint is an optional field, and its implicit value is "False".

With this commit, when an image has its checkpoint property set to "False",
we do not print it in the resulting JSON. Instead, we rely on its implicit
default value.

This has no practical effect and is used as a preparatory commit for the
revise of the checkpoint system.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>