Revise the checkpoint system
In order to compare different checkpoints in different variants we
introduce numbered checkpoints by introducing the flags
"introduces_checkpoint" and "requires_checkpoint".

In this way we can safely signal what is the breaking change that a
particular checkpoint is introducing, and if an image has the
pre-requisites that would allow it to hop between branches.

Having numbered checkpoints also allow us to provide separate generic
fallback and penultimate options. This works around the previous
limitation where the penultimate update was not offered when checkpoints
were involved.

The older "checkpoint" property has been removed. We don't need to
keep the old checkpoint system around because in Jupiter we never
released a checkpoint so far. Additionally, the clients were not parsing
the `checkpoint` property in the meta JSON file that they receive from
the server (which was also an optional field).

The fake_image() function for the automated tests has been tweaked to
accomodate the new parameters. However, we expect to revisit the whole
manifest creation process with a future commit.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>