Add test for a checkpoint that is not incremental
This scenario tests what happen when we release an image that introduces
a particular checkpoint but it requires something different from
`introduces_checkpoint - 1`.

Effectively this is a checkpoint that let you jump from what it
requires, directly to what it introduces, without all the intermediary
steps.

Please note that the `steamdeck-beta/20231113.100.json` case wrongly
tell the client to be up to date. In that scenario we would have
expected an update to `2023113.101` from the same variant
`steamdeck-beta`. This will be handled in the following commit.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>