Search for updates in the same variants as a fallback
When getting the list of all possible candidates, we may have a list
that includes images from different variants from the one we actually
requested. E.g. this may happen when the server is configured with the
`VariantOrder` option.

If in that situation we can't find a suitable update, we should retry
with only candidates that are specifically from the requested variant.
For example when checkpoints are involved, we might find that a specific
image cannot go directly to a more stable variant, but it can apply an
update for the same variant.`

This fixes the automated test included in
`static_rel_checkpoints_expected`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>