Add concept of shadow checkpoints
Shadow checkpoints are not real images and the server will not propose
them as valid candidates.
In this regard they are similar to "skip" images but with a key
difference. Even if they are not proposed as candidates, they still
increase the checkpoint number as if the client actually passed that
checkpoint.

For example in a development branch we could tag two checkpoints that
essentially cancel each other. This could happen if the first checkpoint
turned out to be unnecessary or simply wrong.
In that case we would still be forced to tag two checkpoints for all the
other branches too, otherwise if you were in that development branch you
would not be allowed to change branch anymore.

With this commit we add the ability to create a single "shadow
checkpoint" instead. It basically tells the server that you can jump
to an image past those checkpoints without having to actually install
them.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>