staticserver: Add leftover check for shadow checkpoint
We don't create meta JSON files for shadow checkpoints. This means that
if we notice leftover files for the same images, this is an indication
that something is wrong.
Likely a developer confusingly attempted to convert a regular image
update into a shadow checkpoint.

When this happen we abort the execution to give developers a chance to
cleanup the images hierarchy.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>