imagepool: Remove the unnecessary checkpoints variable
Keeping two separated lists for the checkpoints and the winners is not
necessary anymore.
The comment "It will be used by a future commit" was referring to the
shadow checkpoint implementation, but in the end that was designed
differently.
So, we can clean up the code by removing the `checkpoints` list.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>