When an image manifest is incorrect, abort instead of "silently" continue
If we have image manifests that are wrong or broken, we should not log
an error and simply exclude those images.
Instead, we should raise an error and abort the entire operation.
Log messages can easily be overlooked, and broken manifests are a clear
indication that something is not right.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>