imagepool: Add sanity checks for shadow checkpoints
By its definition, a shadow checkpoint must introduce a checkpoint. If
it don't, either it was not supposed to be shadow checkpoint or a
developer forgot to set the appropriate field `introduces_checkpoint`.

Additionally, a shadow checkpoint can't also be marked as "skip". This
would not make sense.

With this commit we add a sanity check for those two cases and raise an
error if they occur.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>