imagepool: Include additional checkpoint corner cases validations
It is unexpected to have, in the same variant, multiple images with the
same `introduces_checkpoint: X` and `shadow_checkpoint: true`, or
multiple images with `introduces_checkpoint: X`, `shadow_checkpoint:
false` and `skip: false`.

When creating an image pool, if we notice this situation, we abort the
operation and raise an error.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>