staticserver: Limit the estimated download size to only recent images
With the constant increase of new images in our pool, the static server
is getting slower and slower, as time passes, to generate all the
metadata.
At the moment the static server for Jupiter is taking about 2-3 minutes
to complete its process.

By far, the most time intensive operation is pre-estimating the download
sizes.

With this commit we limit the amount of images for which we perform the
download estimation.
We chose an arbitrary number of 150 images as a compromise between
included images and server speed.
To put this number into prospective, in the past 2 years we released
~280 images for Jupiter.
If we continue at this rate, it means about 1 year of updates should be
covered. Instead, if your device is running an image that is older than
that, `steamos-atomupd-client` will perform the size estimation at
runtime on the device itself. That operation usually takes up to 10
seconds to complete.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>