image: Convert values to the correct type when creating an Image
When we use the old dynamic server, we attempt to create an `Image` from
the request arguments provided by the clients.

However, contrary to the expectations, Flask treats all arguments as
strings by default.
This might lead to unexpected situations where, for example, a user
provided argument `checkpoint=False` gets converted into the string
`"False"`. And in Python any non empty strings are evaluated as `True`.

If the args in our dictionary are strings, we convert them into their
expected types.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>