Check if we have images with the same version, release and buildid
Raise an error if in the image pool we have multiple images with the
same version, release and buildid.

If that situation happens, the server cannot reliably identify which one
is the actual image that the client is using.

Additionally, this reduces the confusion for users and developers too,
because having multiple images with the same version and buildid is
not an expected situation.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>