Refactor the meta JSON download from the rest URL
This commit doesn't provide any functional change.

download_update_from_rest_url() has been refactored to make it clearer
what is the generic fallback URL that we try.

The previous implementation was using a `tries` counter because the
initial design allowed more than one generic fallback URL. But the
current design has just one, so there is no need to use a while cycle
and edit the URL in place.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>