image: Include the image checkpoint number in the fallback URL
The static server creates generic fallback meta JSON that include the
checkpoint number information.

When a client needs to use the fallback URL, it should use the
appropriate meta JSON based on the image it is currently running.

If a client is past a checkpoint, the generic `<variant>.json` might
point to an update that the client is not actually able to install.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>